# Improve performance of GenericAffExpr