Fixes a long-standing performance issue in `AffExpr`. 

For now this just accesses a single function from OrderedCollections. I think it's safe to call. The code hasn't changed in 6 years:
https://github.com/JuliaCollections/OrderedCollections.jl/blame/adb9415570faeba7e8c6d800cd743de06104e97e/src/ordered_dict.jl#L320-L337
There's a `_setindex!` function we could call if we wanted to skip the second check,  but that seems like delving into the internals. Another option is to upstream a function, but then we'd need to have a lower bound on OrderedCollections, which I don't know if we want.

## Script

The benchmark script is a bit arbitrary. Creating an expression with unique keys is the same price as before. The benefit comes if there are lots of duplicates in the expression:

```Julia
using BenchmarkTools
using JuMP
N = 1_000
model = Model()
@variable(model, x[1:100])
@benchmark @expression($model, sum(mod(i, 2) * $x[1+mod(i, 100)] for i in 1:$N))
```

## Before

```julia
julia> @benchmark @expression($model, sum(mod(i, 2) * $x[1+mod(i, 100)] for i in 1:$N))
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  24.011 μs …  2.450 ms  ┊ GC (min … max): 0.00% … 98.71%
 Time  (median):     25.276 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   26.805 μs ± 25.057 μs  ┊ GC (mean ± σ):  0.90% ±  0.99%

  ▇██▆▃▁▁▁                                                    ▂
  █████████▇▅▅▄▄▄▃▄▅▄▃▄▆▇██▇▆▆▆▇█▇▇▆▄▄▄▄▃▃▁▅▄▅▅▅▃▅▃▄▅▆▇▇▅▄▁▅▆ █
  24 μs        Histogram: log(frequency) by time      62.8 μs <

 Memory estimate: 5.14 KiB, allocs estimate: 18.
```

## After

```Julia
julia> @benchmark @expression($model, sum(mod(i, 2) * $x[1+mod(i, 100)] for i in 1:$N))
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  15.194 μs … 84.515 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     15.888 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   16.501 μs ±  3.133 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▅██▇▅▃▂▁▁▁▁ ▁▂▁▂▂▁                                          ▂
  ███████████████████▇▆▅▄▅▃▃▃▃▄▁▁▁▁▁▄▁▃▁▄▃▁▃▃▅▃▄▃▃▅▅▄▄▄▅▃▅▄▄▅ █
  15.2 μs      Histogram: log(frequency) by time      32.6 μs <

 Memory estimate: 5.14 KiB, allocs estimate: 18.
```